Struct google_securitycenter1::OrganizationMethods

pub struct OrganizationMethods<'a, C, A> where
    C: 'a,
    A: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on organization resources. It is not used directly, but through the SecurityCommandCenter hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_securitycenter1 as securitycenter1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use securitycenter1::SecurityCommandCenter;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())),
                              <MemoryStorage as Default>::default(), None);
let mut hub = SecurityCommandCenter::new(hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `assets_group(...)`, `assets_list(...)`, `assets_run_discovery(...)`, `assets_update_security_marks(...)`, `get_organization_settings(...)`, `operations_cancel(...)`, `operations_delete(...)`, `operations_get(...)`, `operations_list(...)`, `sources_create(...)`, `sources_findings_create(...)`, `sources_findings_group(...)`, `sources_findings_list(...)`, `sources_findings_patch(...)`, `sources_findings_set_state(...)`, `sources_findings_update_security_marks(...)`, `sources_get(...)`, `sources_get_iam_policy(...)`, `sources_list(...)`, `sources_patch(...)`, `sources_set_iam_policy(...)`, `sources_test_iam_permissions(...)` and `update_organization_settings(...)`
// to build up your call.
let rb = hub.organizations();

Methods

impl<'a, C, A> OrganizationMethods<'a, C, A>

pub fn get_organization_settings(
&self,
name: &str
) -> OrganizationGetOrganizationSettingCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the settings for an organization.

Arguments

name - Name of the organization to get organization settings for. Its format is "organizations/[organization_id]/organizationSettings".

pub fn sources_findings_group(
&self,
request: GroupFindingsRequest,
parent: &str
) -> OrganizationSourceFindingGroupCall<'a, C, A>

Create a builder to help you perform the following task:

Filters an organization or source's findings and groups them by their specified properties.

To group across all sources provide a - as the source id. Example: /v1/organizations/123/sources/-/findings

Arguments

request - No description provided.

parent - Name of the source to groupBy. Its format is "organizations/[organization_id]/sources/[source_id]". To groupBy across all sources provide a source_id of -. For example: organizations/123/sources/-

pub fn sources_list(&self, parent: &str) -> OrganizationSourceListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists all sources belonging to an organization.

Arguments

parent - Resource name of the parent of sources to list. Its format should be "organizations/[organization_id]".

pub fn update_organization_settings(
&self,
request: OrganizationSettings,
name: &str
) -> OrganizationUpdateOrganizationSettingCall<'a, C, A>

Create a builder to help you perform the following task:

Updates an organization's settings.

Arguments

request - No description provided.

name - The relative resource name of the settings. See: https://cloud.google.com/apis/design/resource_names#relative_resource_name Example: "organizations/123/organizationSettings".

pub fn sources_test_iam_permissions(
&self,
request: TestIamPermissionsRequest,
resource: &str
) -> OrganizationSourceTestIamPermissionCall<'a, C, A>

Create a builder to help you perform the following task:

Returns the permissions that a caller has on the specified source.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy detail is being requested. See the operation documentation for the appropriate value for this field.

pub fn assets_run_discovery(
&self,
request: RunAssetDiscoveryRequest,
parent: &str
) -> OrganizationAssetRunDiscoveryCall<'a, C, A>

Create a builder to help you perform the following task:

Runs asset discovery. The discovery is tracked with a long-running operation.

This API can only be called with limited frequency for an organization. If it is called too frequently the caller will receive a TOO_MANY_REQUESTS error.

Arguments

request - No description provided.

parent - Name of the organization to run asset discovery for. Its format is "organizations/[organization_id]".

pub fn sources_get_iam_policy(
&self,
request: GetIamPolicyRequest,
resource: &str
) -> OrganizationSourceGetIamPolicyCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the access control policy on the specified Source.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy is being requested. See the operation documentation for the appropriate value for this field.

pub fn operations_cancel(
&self,
name: &str
) -> OrganizationOperationCancelCall<'a, C, A>

Create a builder to help you perform the following task:

Starts asynchronous cancellation on a long-running operation. The server makes a best effort to cancel the operation, but success is not guaranteed. If the server doesn't support this method, it returns google.rpc.Code.UNIMPLEMENTED. Clients can use Operations.GetOperation or other methods to check whether the cancellation succeeded or whether the operation completed despite cancellation. On successful cancellation, the operation is not deleted; instead, it becomes an operation with an Operation.error value with a google.rpc.Status.code of 1, corresponding to Code.CANCELLED.

Arguments

name - The name of the operation resource to be cancelled.

pub fn sources_findings_set_state(
&self,
request: SetFindingStateRequest,
name: &str
) -> OrganizationSourceFindingSetStateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates the state of a finding.

Arguments

request - No description provided.

name - The relative resource name of the finding. See: https://cloud.google.com/apis/design/resource_names#relative_resource_name Example: "organizations/123/sources/456/finding/789".

pub fn sources_findings_list(
&self,
parent: &str
) -> OrganizationSourceFindingListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists an organization or source's findings.

To list across all sources provide a - as the source id. Example: /v1/organizations/123/sources/-/findings

Arguments

parent - Name of the source the findings belong to. Its format is "organizations/[organization_id]/sources/[source_id]". To list across all sources provide a source_id of -. For example: organizations/123/sources/-

pub fn sources_findings_patch(
&self,
request: Finding,
name: &str
) -> OrganizationSourceFindingPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Creates or updates a finding. The corresponding source must exist for a finding creation to succeed.

Arguments

request - No description provided.

name - The relative resource name of this finding. See: https://cloud.google.com/apis/design/resource_names#relative_resource_name Example: "organizations/123/sources/456/findings/789"

pub fn assets_group(
&self,
request: GroupAssetsRequest,
parent: &str
) -> OrganizationAssetGroupCall<'a, C, A>

Create a builder to help you perform the following task:

Filters an organization's assets and groups them by their specified properties.

Arguments

request - No description provided.

parent - Name of the organization to groupBy. Its format is "organizations/[organization_id]".

pub fn sources_findings_update_security_marks(
&self,
request: SecurityMarks,
name: &str
) -> OrganizationSourceFindingUpdateSecurityMarkCall<'a, C, A>

Create a builder to help you perform the following task:

Updates security marks.

Arguments

request - No description provided.

name - The relative resource name of the SecurityMarks. See: https://cloud.google.com/apis/design/resource_names#relative_resource_name Examples: "organizations/123/assets/456/securityMarks" "organizations/123/sources/456/findings/789/securityMarks".

pub fn sources_patch(
&self,
request: Source,
name: &str
) -> OrganizationSourcePatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates a source.

Arguments

request - No description provided.

name - The relative resource name of this source. See: https://cloud.google.com/apis/design/resource_names#relative_resource_name Example: "organizations/123/sources/456"

pub fn assets_list(&self, parent: &str) -> OrganizationAssetListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists an organization's assets.

Arguments

parent - Name of the organization assets should belong to. Its format is "organizations/[organization_id]".

pub fn operations_get(
&self,
name: &str
) -> OrganizationOperationGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the latest state of a long-running operation. Clients can use this method to poll the operation result at intervals as recommended by the API service.

Arguments

name - The name of the operation resource.

pub fn operations_list(
&self,
name: &str
) -> OrganizationOperationListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists operations that match the specified filter in the request. If the server doesn't support this method, it returns UNIMPLEMENTED.

NOTE: the name binding allows API services to override the binding to use different resource name schemes, such as users/*/operations. To override the binding, API services can add a binding such as "/v1/{name=users/*}/operations" to their service configuration. For backwards compatibility, the default name includes the operations collection id, however overriding users must ensure the name binding is the parent resource, without the operations collection id.

Arguments

name - The name of the operation's parent resource.

pub fn assets_update_security_marks(
&self,
request: SecurityMarks,
name: &str
) -> OrganizationAssetUpdateSecurityMarkCall<'a, C, A>

Create a builder to help you perform the following task:

Updates security marks.

Arguments

request - No description provided.

name - The relative resource name of the SecurityMarks. See: https://cloud.google.com/apis/design/resource_names#relative_resource_name Examples: "organizations/123/assets/456/securityMarks" "organizations/123/sources/456/findings/789/securityMarks".

pub fn sources_set_iam_policy(
&self,
request: SetIamPolicyRequest,
resource: &str
) -> OrganizationSourceSetIamPolicyCall<'a, C, A>

Create a builder to help you perform the following task:

Sets the access control policy on the specified Source.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy is being specified. See the operation documentation for the appropriate value for this field.

pub fn sources_create(
&self,
request: Source,
parent: &str
) -> OrganizationSourceCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a source.

Arguments

request - No description provided.

parent - Resource name of the new source's parent. Its format should be "organizations/[organization_id]".

pub fn sources_findings_create(
&self,
request: Finding,
parent: &str
) -> OrganizationSourceFindingCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a finding. The corresponding source must exist for finding creation to succeed.

Arguments

request - No description provided.

parent - Resource name of the new finding's parent. Its format should be "organizations/[organization_id]/sources/[source_id]".

pub fn operations_delete(
&self,
name: &str
) -> OrganizationOperationDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a long-running operation. This method indicates that the client is no longer interested in the operation result. It does not cancel the operation. If the server doesn't support this method, it returns google.rpc.Code.UNIMPLEMENTED.

Arguments

name - The name of the operation resource to be deleted.

pub fn sources_get(&self, name: &str) -> OrganizationSourceGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets a source.

Arguments

name - Relative resource name of the source. Its format is "organizations/[organization_id]/source/[source_id]".